# backend/docs/openapi-governance.yaml
openapi: 3.1.0
info:
  title: DictionaryRoot Governed Platform API
  version: 0.2.0
  description: Authentication, account, governance, workflow, moderation, and audit additions to SourceRoot.
servers:
  - url: http://localhost:3000/api/v1
paths:
...
  /dictionaryroot/workflow/proposals:
    get:
      summary: List visible change proposals
      responses: { '200': { description: Proposal queue } }
    post:
      summary: Create a sourced change proposal
      responses: { '201': { description: Proposal created }, '403': { description: Permission required } }
...
  /dictionaryroot/workflow/proposals/{proposalId}/publish:
    post:
      summary: Publish an approved, unlocked proposal
      parameters:
        - in: path
          name: proposalId
          required: true
          schema: { type: string, format: uuid }
      responses: { '201': { description: Publication created }, '423': { description: Target is moderation locked } }
  /dictionaryroot/workflow/publications/{publicationId}/rollback:
    post:
      summary: Roll back a publication with a recorded reason
      parameters:
        - in: path
          name: publicationId
          required: true
          schema: { type: string, format: uuid }
      responses: { '200': { description: Publication rolled back } }
...
components:
  securitySchemes:
    dictionaryRootSession:
      type: apiKey
      in: cookie
      name: dr_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token